+++ src/AudioSync.OnsetDetection/Util/SpectralDescription/BaseSpectralDescription.cs
@@ -4,5 +4,5 @@ namespace AudioSync.OnsetDetection.Util.SpectralDescription;
 
 internal abstract class BaseSpectralDescription
 {
-    public abstract void Perform(in Span<Polar> fftGrain, ref double onset);
+    public abstract void Perform(in Span<Polar> fftGrain, ref float onset);
 }
