+++ src/AudioSync.OnsetDetection/Util/SpectralDescription/ComplexDomain.cs
@@ -4,20 +4,20 @@ namespace AudioSync.OnsetDetection.Util.SpectralDescription;
 
 internal sealed class ComplexDomain : BaseSpectralDescription
 {
-    private readonly double[] oldMagnitude;
-    private readonly double[] currentMeasureVector;
-    private readonly double[] lastFrameVector;
-    private readonly double[] secondLastFrameVector;
+    private readonly float[] oldMagnitude;
+    private readonly float[] currentMeasureVector;
+    private readonly float[] lastFrameVector;
+    private readonly float[] secondLastFrameVector;
 
     public ComplexDomain(int realSize)
     {
-        oldMagnitude = new double[realSize];
-        currentMeasureVector = new double[realSize];
-        lastFrameVector = new double[realSize];
-        secondLastFrameVector = new double[realSize];
+        oldMagnitude = new float[realSize];
+        currentMeasureVector = new float[realSize];
+        lastFrameVector = new float[realSize];
+        secondLastFrameVector = new float[realSize];
     }
 
-    public override void Perform(in Span<Polar> fftGrain, ref double onset)
+    public override void Perform(in Span<Polar> fftGrain, ref float onset)
     {
         var length = fftGrain.Length;
         onset = default;
@@ -28,13 +28,13 @@ internal sealed class ComplexDomain : BaseSpectralDescription
             ref var grain = ref fftGrain[i];
 
             // compute predicted phase
-            currentMeasureVector[i] = (2.0 * lastFrameVector[i]) - secondLastFrameVector[i];
+            currentMeasureVector[i] = (2.0f * lastFrameVector[i]) - secondLastFrameVector[i];
 
             // compute euclidean distance in the complex domain
             var euclideanDistance = (oldMag * oldMag) + (grain.Norm * grain.Norm);
-            euclideanDistance -= 2 * oldMag * grain.Norm * Math.Cos(currentMeasureVector[i] - grain.Phase);
-            euclideanDistance = Math.Abs(euclideanDistance);
-            euclideanDistance = Math.Sqrt(euclideanDistance);
+            euclideanDistance -= 2 * oldMag * grain.Norm * (float)Math.Cos(currentMeasureVector[i] - grain.Phase);
+            euclideanDistance = (float)Math.Abs(euclideanDistance);
+            euclideanDistance = (float)Math.Sqrt(euclideanDistance);
             onset += euclideanDistance;
 
             // Push back frames
