+++ src/AudioSync.Tests/MusicTests.cs
@@ -6,20 +6,20 @@ namespace AudioSync.Tests;
 // All test songs *MUST* use ogg!
 // NAudio mp3 support is Windows exclusive.
 
-[TestFixture("Music/Creo/Place on Fire.ogg", 70)] // Creo - Place on Fire @ 70 BPM
-[TestFixture("Music/Creo/Rock Thing.ogg", 114.0)] // Creo - Rock Thing @ 114 BPM
-[TestFixture("Music/Creo/Showdown.ogg", 110.0)] // Creo - Showdown @ 110 BPM
-[TestFixture("Music/Creo/Ahead of the Curve.ogg", 96.0)] // Creo - Rock Thing @ 96 BPM
-[TestFixture("Music/Creo/Sphere.ogg", 100.0)] // Creo - Sphere @ 100 BPM
-[TestFixture("Music/Creo/Dimension.ogg", 115.0)] // Creo - Dimension @ 115 BPM
-[TestFixture("Music/Creo/322.ogg", 128.0)] // Creo - 322 @ 128 BPM
+[TestFixture("Music/Creo/Place on Fire.ogg", 70f)] // Creo - Place on Fire @ 70 BPM
+[TestFixture("Music/Creo/Rock Thing.ogg", 114.0f)] // Creo - Rock Thing @ 114 BPM
+[TestFixture("Music/Creo/Showdown.ogg", 110.0f)] // Creo - Showdown @ 110 BPM
+[TestFixture("Music/Creo/Ahead of the Curve.ogg", 96.0f)] // Creo - Rock Thing @ 96 BPM
+[TestFixture("Music/Creo/Sphere.ogg", 100.0f)] // Creo - Sphere @ 100 BPM
+[TestFixture("Music/Creo/Dimension.ogg", 115.0f)] // Creo - Dimension @ 115 BPM
+[TestFixture("Music/Creo/322.ogg", 128.0f)] // Creo - 322 @ 128 BPM
 public class MusicTests
 {
     private readonly SyncAnalyser syncAnalyser;
     private readonly string song;
-    private readonly double expectedBPM;
+    private readonly float expectedBPM;
 
-    public MusicTests(string song, double expectedBPM)
+    public MusicTests(string song, float expectedBPM)
     {
         syncAnalyser = new();
