+++ src/AudioSync.Util.Tests/Multiply/MultiplyTest.cs
@@ -7,9 +7,9 @@ public class MultiplyTest
     [Test]
     public void SimpleMultiplyTest()
     {
-        Span<double> left = stackalloc double[] { 1, 2, 3 }; 
-        Span<double> right = stackalloc double[] { 2, 2, 2 };
-        Span<double> result = stackalloc double[3];
+        Span<float> left = stackalloc float[] { 1, 2, 3 }; 
+        Span<float> right = stackalloc float[] { 2, 2, 2 };
+        Span<float> result = stackalloc float[3];
 
         Utils.Multiply(in left, in right, ref result);
 
@@ -24,9 +24,9 @@ public class MultiplyTest
     {
         Assert.Throws<AudioSyncFatalException>(() =>
         {
-            Span<double> left = stackalloc double[10];
-            Span<double> right = stackalloc double[5];
-            Span<double> result = stackalloc double[10];
+            Span<float> left = stackalloc float[10];
+            Span<float> right = stackalloc float[5];
+            Span<float> result = stackalloc float[10];
 
             // Should throw because the length of our left and right operands are not equal.
             Utils.Multiply(in left, in right, ref result);
@@ -38,9 +38,9 @@ public class MultiplyTest
     {
         Assert.Throws<AudioSyncFatalException>(() =>
         {
-            Span<double> left = stackalloc double[10];
-            Span<double> right = stackalloc double[10];
-            Span<double> result = stackalloc double[5];
+            Span<float> left = stackalloc float[10];
+            Span<float> right = stackalloc float[10];
+            Span<float> result = stackalloc float[5];
 
             // Should throw because the length of our result operand isnt equal to the operands
             Utils.Multiply(in left, in right, ref result);
@@ -52,9 +52,9 @@ public class MultiplyTest
     {
         Assert.Throws<AudioSyncFatalException>(() =>
         {
-            Span<double> left = stackalloc double[10];
-            Span<double> right = stackalloc double[5];
-            Span<double> result = stackalloc double[10];
+            Span<float> left = stackalloc float[10];
+            Span<float> right = stackalloc float[5];
+            Span<float> result = stackalloc float[10];
 
             // Should throw because the length of our left and right operands are not equal.
             Utils.MultiplySIMD(in left, in right, ref result);
@@ -66,9 +66,9 @@ public class MultiplyTest
     {
         Assert.Throws<AudioSyncFatalException>(() =>
         {
-            Span<double> left = stackalloc double[10];
-            Span<double> right = stackalloc double[10];
-            Span<double> result = stackalloc double[5];
+            Span<float> left = stackalloc float[10];
+            Span<float> right = stackalloc float[10];
+            Span<float> result = stackalloc float[5];
 
             // Should throw because the length of our result operand isnt equal to the operands
             Utils.MultiplySIMD(in left, in right, ref result);
